managers/ec2: starting to refactor the instance creation
Now it will be done in background, by another process.

Here is the flow:

  1. service-add will create the instance in the database, with
     the "creating" status
  2. a process will get all instances in "creating" status and call a
     method in the manager to create the "physical" instance (something
     like scale_instance(instance_name, 1)), and then change the state
     to "created"

In order to make sure that the process described in the second step is
parallel safe, we will use a distributed lock to change the state to
"starting". So we can have as much instances of this process as we want.
In pseudocode:

  locker.lock("instance_starting")
  instance = collection.find_one({state: "creating"})
  collection.update({name: instance.name}, {$set: {state: "starting"}})
  locker.unlock("instance_starting")
  # do things with the instance

The next step is providing this process (somethig similar to the VCL
writer) and a way of running it (I will include it in the sample
Procfile).